Standardized the Firebolt selector to only return a NodeCollection
This greatly reduces the selector's performance, but now you always know what is being returned when you call the function.
Every effort has been made to maximize the selector's performance when selecting elements in the current document. This includes using function binding to improve performance in certain browsers, which also improves the performance of the other global selector functions.